Refactor ContractEvents to make it easier to integrate graph
